src/ai-service/requirements.txt
fastapi==0.115.1
uvicorn==0.34.0
pydantic==2.10.6
pytest==8.3.5
httpx==0.28.1
pyyaml==6.0.2
semantic-kernel==1.24.1
azure.identity==1.21.0
requests==2.32.3
openai==1.66.3
pillow==11.1.0